Project/requirements.txt
Flask
requests
beautifulsoup4
PyMuPDF
google-generativeai
gunicorn